Change `int` to `uint16_t` also in obstacle.hpp
I missed that file. It wouldn't do much to do it, but having about 4 million different id, which can't be used anyway, because the map can take only 2^16 different id doesn't make sense.

+++ src/map/obstacle.hpp
@@ -29,7 +29,7 @@ public:
 		this->id = obstacle.get_id();
 	}
 
-	Obstacle(int id, bool collision = false, bool shadow = false) {
+	Obstacle(uint16_t id, bool collision = false, bool shadow = false) {
 
 		this->collision = collision;
 		this->shadow = shadow;
@@ -43,7 +43,7 @@ public:
 	/**
      *  Return obstacle's id
      */
-	inline int get_id() { return id; }
+	inline uint16_t get_id() { return id; }
 
 	/**
      *  Return obstacle's collision state
